Move buildConditions to a common helper
